BACKUP_20251227/Financial Engineering API Demo/requirements.txt
# Core Dependencies
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# API Clients
requests>=2.31.0
yfinance>=0.2.28
alpha-vantage>=2.3.1
PyGithub>=1.59.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Data Processing
python-dotenv>=1.0.0
cachetools>=5.3.0

# Financial Calculations
quantlib>=1.31
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0
